Refactor WoW64 Engine to utilize existing CapeMon hook routines
- Extracted hook_create_trampoline and assembly pointer displacement correction logic into trampoline_64.c, allowing both capemon_x64 and capemon_wow64 Native environments to share the exact same disassembler logic without code duplication.
- Migrated internal _DInst fetching inside trampoline generation away from malloc to support /NODEFAULTLIB CRT-less capabilities required by WOW64 mitigations.
- Replaced handwritten hook templates in wow64_engine with HOOKDEF macros matching CapeMon's native api framework formats.
