+++ capemon.vcxproj.filters
@@ -183,7 +183,7 @@
     <ClCompile Include="hooking_32.c">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="hooking_64.c">
+    <ClCompile Include="hooking_64.c">\n    </ClCompile>\n    <ClCompile Include="trampoline_64.c">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="hooking.c">
